move convert function to utils.py
